HW2 bigChef AC

+++ HW2_BigChefDanBin/bigChef.cpp
@@ -7,10 +7,10 @@ using namespace std;
 
 struct People
 {
-    int a;
-    int b;
+    long long int a;
+    long long int b;
     People() = default;
-    People(int A, int B){
+    People(long long int A, long long int B){
         a = A;
         b = B;
     }
@@ -19,12 +19,12 @@ struct People
 class Group
 {
 public:
-    int m; // m people
-    int max_eachGroup_CE = 0;
-    int max_eachGroup_C = 0;
+    long long int m; // m people
+    long long int max_eachGroup_CE = 0;
+    long long int max_eachGroup_C = 0;
     People * peoples;
     Group() = default;
-    Group(int M): m(M) {
+    Group(long long int M): m(M) {
         peoples = new People[m];
     }
     ~Group(){};
@@ -44,17 +44,17 @@ bool waitTime__(Group g1, Group g2)
     return false;
 }
 
-int cooking(int N, Group* groupArr)
+long long int cooking(long long int N, Group* groupArr)
 {
     // 吃的最久的人先吃，先把每組依吃的時間排序
-    for(int i = 0; i< N; i++)
+    for(long long int i = 0; i< N; i++)
     {
         sort(groupArr[i].peoples, groupArr[i].peoples + groupArr[i].m, eating__);
         // 計算最大吃+ 煮的時間
-        groupArr[i].max_eachGroup_CE = INT32_MIN;
+        groupArr[i].max_eachGroup_CE = LLONG_MIN;
         // 計算最大煮的時間
         groupArr[i].max_eachGroup_C = 0;
-        for(int j = 0; j< groupArr[i].m; j++)
+        for(long long int j = 0; j< groupArr[i].m; j++)
         {
             groupArr[i].max_eachGroup_C += groupArr[i].peoples[j].a;
             if(groupArr[i].max_eachGroup_C + groupArr[i].peoples[j].b > groupArr[i].max_eachGroup_CE)
@@ -65,16 +65,16 @@ int cooking(int N, Group* groupArr)
     // Merge sort 排序讓別人等的時間少的排在前面
     sort(groupArr, groupArr + N, waitTime__);
 
-    int current_cooking_time=0;
-    for(int i = 0; i< N-1; i++)
+    long long int current_cooking_time=0;
+    for(long long int i = 0; i< N-1; i++)
     {
         current_cooking_time += groupArr[i].max_eachGroup_C;
         groupArr[i+1].max_eachGroup_CE += current_cooking_time;
         
     }
 
-    int count  = 0;
-    for(int i = 0; i< N; i++)
+    long long int count  = 0;
+    for(long long int i = 0; i< N; i++)
     {
         //cout << groupArr[i].max_eachGroup_CE << "  ";
         count += groupArr[i].max_eachGroup_CE * groupArr[i].m;
@@ -88,15 +88,15 @@ int main()
 {
     std::ios_base::sync_with_stdio(false);
     std::cin.tie(nullptr);
-    int N, M, tempA, tempB;
+    long long int N, M, tempA, tempB;
     cin >> N;
     Group* groupArr = new Group[N];
-    for(int i = 0; i< N; i++)
+    for(long long int i = 0; i< N; i++)
     {
         cin >> M;
         groupArr[i] = Group(M);
         
-        for(int j = 0; j < M; j++)
+        for(long long int j = 0; j < M; j++)
         {
             cin >> tempA >> tempB;
             groupArr[i].peoples[j] = People(tempA, tempB);
